feat(webhook): Refactor webhoko events subscriptions
